fix: 修复主题/设置闪退、搜索不可用、阅读器状态栏透视问题
- 修复 Color(ULong) 构造函数误用导致主题/字体/设置按钮闪退
- 修复默认书源仅在书源管理页加载，移至 Application 启动时加载
- 移除搜索健康检查过滤，修复健康检查用中文测英文源的误判
- 阅读器顶部添加状态栏背景遮罩，防止内容透过状态栏可见
- 默认字体大小从 18 改为 24

+++ app/src/main/java/com/linghualive/flamekit/core/theme/ReaderTheme.kt
@@ -46,9 +46,9 @@ fun readerColorsFor(themeType: ReaderThemeType, customTheme: CustomReaderTheme? 
         val theme = customTheme ?: CustomReaderTheme()
         try {
             ReaderColors(
-                background = Color(theme.backgroundColor.toULong()),
-                textColor = Color(theme.textColor.toULong()),
-                secondaryText = Color(theme.secondaryTextColor.toULong()),
+                background = Color(theme.backgroundColor.toInt()),
+                textColor = Color(theme.textColor.toInt()),
+                secondaryText = Color(theme.secondaryTextColor.toInt()),
             )
         } catch (_: Exception) {
             // Fallback to light theme if custom color values are invalid
